CRUD Employee

+++ src/app/components/employee/employee.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { EmployeeService } from '../../services/employee.service';
 import { AuthService } from '../../services/auth.service';
 import { Employee } from '../../model/employee.model';
+import swal from 'sweetalert2'
 
 declare var $: any; // jQuery
 
@@ -25,6 +26,11 @@ export class EmployeeComponent implements OnInit {
     this.getAllEmployee();
   }
 
+  createNew() {
+    this.employee = new Employee();
+    this.showModal();
+  }
+
   getAllEmployee() {
     this.employeeService.getAllEmployee()
       .subscribe(
@@ -32,6 +38,16 @@ export class EmployeeComponent implements OnInit {
           if (employees != null) {
             this.employees = employees;
           }
+        }
+      );
+  }
+
+  viewEmployee(id: any) {
+    this.employeeService.getEmployee(id)
+      .subscribe(
+        (employee: Employee) => {
+          this.employee = employee;
+          this.showModal();
         }
       );
   }
@@ -44,28 +60,65 @@ export class EmployeeComponent implements OnInit {
     $('#employee-modal').modal('hide');
   }
 
-  onSubmit(employee: Employee) {
-    this.employeeService.createEmployee(employee)
-      .subscribe(
-        (res: any) => {
-          if (res.result === 'success') {
-            console.log(res.message);
-            this.getAllEmployee();
-            this.hideModal();
+  onSubmit() {
+    if (this.employee.id) {
+      this.employeeService.updateEmployee(this.employee)
+        .subscribe(
+          (res: any) => {
+            if (res.result === 'success') {
+              this.getAllEmployee();
+              swal(
+                'Success!',
+                'Employee Information successfully updated!',
+                'success'
+              );
+              this.hideModal();
+            }
           }
-        }
-      );
+        );
+    } else {
+      this.employeeService.createEmployee(this.employee)
+        .subscribe(
+          (res: any) => {
+            if (res.result === 'success') {
+              this.getAllEmployee();
+              swal(
+                'Success!',
+                'New Employee has been Added!',
+                'success'
+              );
+              this.hideModal();
+            }
+          }
+        );
+    }
   }
 
   deleteEmployee(id: any) {
-    this.employeeService.deleteEmployee(id)
-      .subscribe(
-        (res: any) => {
-          if (res.result === 'success') {
-            this.getAllEmployee();
-          }
-        }
-      );
+    swal({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      type: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.value) {
+        this.employeeService.deleteEmployee(id)
+          .subscribe(
+            (res: any) => {
+              if (res.result === 'success') {
+                swal(
+                  'Deleted!',
+                  'Employee Information has been deleted.',
+                  'success'
+                );
+                this.getAllEmployee();
+              }
+            }
+          );
+      }
+    });
   }
-
 }
